+++ src/main/java/org/grnet/pidmr/repository/RoleChangeRequestsRepository.java
@@ -27,6 +27,7 @@ public class RoleChangeRequestsRepository implements Repository<RoleChangeReques
 
         return find("userId = : userId", Parameters.with("userId", userId)).stream().findAny();
     }
+
     /**
      * Retrieves a page of from the database.
      *
@@ -47,5 +48,25 @@ public class RoleChangeRequestsRepository implements Repository<RoleChangeReques
         return pageable;
     }
 
+    /**
+     * Finds role change requests by user ID with pagination.
+     *
+     * @param userId the ID of the user whose role change requests are to be retrieved
+     * @param page The index of the page to retrieve (starting from 0).
+     * @param size The maximum number of Providers to include in a page.
+     * @return a paginated list of role change requests
+     */
+    public Pageable<RoleChangeRequest> fetchRoleChangeRequestByUser(int page, int size, String userId){
+
+        var panache = find("from RoleChangeRequest r where r.userId = ?1", Sort.by("status", Sort.Direction.Ascending).and("requestedOn", Sort.Direction.Descending), userId).page(page, size);
+        var pageable = new PageableImpl<RoleChangeRequest>();
+        pageable.list = panache.list();
+        pageable.index = page;
+        pageable.size = size;
+        pageable.count = panache.count();
+        pageable.page = Page.of(page, size);
+
+        return pageable;
+    }
 
 }